+++ PrintLogUploader.py
@@ -4,11 +4,24 @@ import time
 import collections
 from typing import Optional, TYPE_CHECKING
 
-from PyQt6.QtQml import qmlRegisterType
-from PyQt6.QtCore import QObject, QBuffer
-from PyQt6.QtNetwork import QNetworkRequest
-from PyQt6.QtWidgets import QMessageBox
-from PyQt6.QtGui import QPixmap
+USE_QT5 = False
+try:
+    from PyQt6.QtQml import qmlRegisterType
+    from PyQt6.QtCore import QObject, QBuffer
+    from PyQt6.QtNetwork import QNetworkRequest
+    from PyQt6.QtWidgets import QMessageBox
+    from PyQt6.QtGui import QPixmap
+    if TYPE_CHECKING:
+        from PyQt6.QtNetwork import QNetworkReply
+except ImportError:
+    from PyQt5.QtQml import qmlRegisterType
+    from PyQt5.QtCore import QObject, QBuffer
+    from PyQt5.QtNetwork import QNetworkRequest
+    from PyQt5.QtWidgets import QMessageBox
+    from PyQt5.QtGui import QPixmap
+    if TYPE_CHECKING:
+        from PyQt5.QtNetwork import QNetworkReply
+    USE_QT5 = True
 
 from cura.CuraApplication import CuraApplication
 
@@ -91,6 +104,11 @@ class PrintLogUploader(QObject, Extension):
             True
         )
 
+        self._application.getPreferences().addPreference(
+            "3d_print_log/bypass_prompt",
+            False
+        )
+
         self._application.engineCreatedSignal.connect(self._onEngineCreated)
 
     def _onSendMenuButtonClicked(self):
@@ -112,8 +130,14 @@ class PrintLogUploader(QObject, Extension):
         )
 
     def showSettingsDialog(self):
-        path = os.path.join(os.path.dirname(
-            os.path.abspath(__file__)), "qml", "SettingsDialog.qml")
+        path = None
+        if USE_QT5:
+            path = os.path.join(os.path.dirname(
+            os.path.abspath(__file__)), "qml", "qt5", "SettingsDialog.qml")
+        else:
+            path = os.path.join(os.path.dirname(
+            os.path.abspath(__file__)), "qml", "qt6","SettingsDialog.qml")
+
         self._settings_dialog = CuraApplication.getInstance(
         ).createQmlComponent(path, {"manager": self})
         self._settings_dialog.show()
@@ -184,11 +208,20 @@ class PrintLogUploader(QObject, Extension):
         if not hasSliced:
             return False
 
+        preferences = self._application.getInstance().getPreferences()
+        bypass_prompt = preferences.getValue(
+                "3d_print_log/bypass_prompt")
+        if (bypass_prompt):
+                return True
+
         dialog = self._createConfirmationDialog()
 
         returnValue = dialog.exec()
-
-        return returnValue == QMessageBox.Ok
+        
+        if USE_QT5:
+            return returnValue == QMessageBox.Ok
+        else:
+            return returnValue == QMessageBox.StandardButton.Ok
 
     def _hasSlicedModel(self) -> bool:
         '''Checks to see if the model has been sliced'''
@@ -205,11 +238,11 @@ class PrintLogUploader(QObject, Extension):
     def _createConfirmationDialog(self):
         '''Create a message box prompting the user if they want to send this print information.'''
         msgBox = QMessageBox()
-        msgBox.setIcon(QMessageBox.Information)
+        msgBox.setIcon(QMessageBox.Information if USE_QT5 else QMessageBox.Icon.Information)
         msgBox.setText("Would you like to send to 3Dprintlog.com?")
         msgBox.setWindowTitle("Send to 3D Print Log?")
-        msgBox.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
-        msgBox.setDefaultButton(QMessageBox.Ok)
+        msgBox.setStandardButtons((QMessageBox.Ok if USE_QT5 else QMessageBox.StandardButton.Ok) | (QMessageBox.Cancel if USE_QT5 else QMessageBox.StandardButton.Cancel))
+        msgBox.setDefaultButton(QMessageBox.Ok if USE_QT5 else QMessageBox.StandardButton.Ok)
 
         self._add3DPrintLogLogo(msgBox)
 
@@ -218,11 +251,11 @@ class PrintLogUploader(QObject, Extension):
     def _createDialog(self, text, title):
         '''Create a messsage box with a title and text'''
         msgBox = QMessageBox()
-        msgBox.setIcon(QMessageBox.Information)
+        msgBox.setIcon(QMessageBox.Information if USE_QT5 else QMessageBox.Icon.Information)
         msgBox.setText(text)
         msgBox.setWindowTitle(title)
-        msgBox.setStandardButtons(QMessageBox.Ok)
-        msgBox.setDefaultButton(QMessageBox.Ok)
+        msgBox.setStandardButtons((QMessageBox.Ok if USE_QT5 else QMessageBox.StandardButton.Ok))
+        msgBox.setDefaultButton((QMessageBox.Ok if USE_QT5 else QMessageBox.StandardButton.Ok))
 
         self._add3DPrintLogLogo(msgBox)
 
@@ -252,7 +285,13 @@ class PrintLogUploader(QObject, Extension):
 
     def _onRequestFinished(self, reply: "QNetworkReply") -> None:
         '''Handle the response from the API after sending the settings.'''
-        status_code = reply.attribute(QNetworkRequest.Attribute.HttpStatusCodeAttribute)
+        status_code = None
+
+        if USE_QT5:
+            status_code = reply.attribute(QNetworkRequest.HttpStatusCodeAttribute)
+        else:
+            status_code = reply.attribute(QNetworkRequest.Attribute.HttpStatusCodeAttribute)
+
         if status_code == 200:
             # API will return a GUID that can be used to retrieve the setting information that was saved.
             results = json.loads(reply.readAll().data().decode("utf-8"))
@@ -314,7 +353,10 @@ class PrintLogUploader(QObject, Extension):
                 Logger.log("i", "Snapshot Found")
 
                 thumbnail_buffer = QBuffer()
-                thumbnail_buffer.open(QBuffer.OpenModeFlag.ReadWrite)
+                if USE_QT5:
+                    thumbnail_buffer.open(QBuffer.ReadWrite)
+                else:
+                    thumbnail_buffer.open(QBuffer.OpenModeFlag.ReadWrite)
                 snapshot.save(thumbnail_buffer, "PNG")
 
                 encodedSnapshot = thumbnail_buffer.data(