+++ __init__.py
@@ -1,7 +1,12 @@
 from . import PrintLogUploader
 from . import PrintLogSettingDefinitionsModel
 
-from PyQt6.QtQml import qmlRegisterType
+USE_QT5 = False
+try:
+    from PyQt6.QtQml import qmlRegisterType
+except ImportError:
+    from PyQt5.QtQml import qmlRegisterType
+    USE_QT5 = True
 
 
 def getMetaData():