+++ qml/qt5/SettingCategory.qml
@@ -1,7 +1,9 @@
+
 import QtQuick 2.2
 import QtQuick.Controls 1.1
 import QtQuick.Controls.Styles 1.1
 import QtQuick.Layouts 1.1
+
 
 import UM 1.1 as UM